Make travis great again, test fixes, pep8 fixes
