+++ requirements.txt
@@ -2,3 +2,5 @@ PyStemmer>=1.3,<2.0
 tensorflow>=1.0,<2.0
 bblfsh>=2.2.1,<3.0
 modelforge>=0.5.0-alpha
+sourced-engine>=0.2.0,<0.3
+parquet>=1.2,<2.0
