+++ sourced/ml/tests/test_id_embedding.py
@@ -11,8 +11,8 @@ from scipy.sparse import coo_matrix
 
 from modelforge.model import split_strings, assemble_sparse_matrix
 from sourced.ml.algorithms import swivel
-from sourced.ml.algorithms.id_embedding import preprocess, run_swivel, postprocess, SwivelTransformer, \
-    PostprocessTransformer, PreprocessTransformer
+from sourced.ml.algorithms.id_embedding import preprocess, run_swivel, postprocess, \
+    SwivelTransformer, PostprocessTransformer, PreprocessTransformer
 from sourced.ml.models import DocumentFrequencies, Id2Vec
 from sourced.ml.tests.test_dump import captured_output
